requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-generativeai>=0.7.0S